refactor: Move uninstall exceptions to their own submodule
