+++ src/pip/_internal/exceptions/__init__.py
@@ -41,6 +41,10 @@ from pip._internal.exceptions.pyproject import (
     InvalidPyProjectBuildRequires,
     MissingPyProjectBuildRequires,
 )
+from pip._internal.exceptions.uninstall import (
+    LegacyDistutilsInstall,
+    UninstallMissingRecord,
+)
 from pip._internal.exceptions.wheel import (
     InvalidWheel,
     InvalidWheelFilename,
@@ -67,6 +71,9 @@ __all__ = [
     # pyproject
     "InvalidPyProjectBuildRequires",
     "MissingPyProjectBuildRequires",
+    # uninstall
+    "LegacyDistutilsInstall",
+    "UninstallMissingRecord",
     # wheel
     "InvalidWheel",
     "InvalidWheelFilename",
@@ -744,48 +751,6 @@ class ExternallyManagedEnvironment(DiagnosticPipError):
             exc_info = logger.isEnabledFor(VERBOSE)
             logger.warning("Failed to read %s", config, exc_info=exc_info)
         return cls(None)
-
-
-class UninstallMissingRecord(DiagnosticPipError):
-    reference = "uninstall-no-record-file"
-
-    def __init__(self, *, distribution: BaseDistribution) -> None:
-        installer = distribution.installer
-        if not installer or installer == "pip":
-            dep = f"{distribution.raw_name}=={distribution.version}"
-            hint = Text.assemble(
-                "You might be able to recover from this via: ",
-                (f"pip install --ignore-installed --no-deps {dep}", "green"),
-            )
-        else:
-            hint = Text(
-                f"The package was installed by {installer}. "
-                "You should check if it can uninstall the package."
-            )
-
-        super().__init__(
-            message=Text(f"Cannot uninstall {distribution}"),
-            context=(
-                "The package's contents are unknown: "
-                f"no RECORD file was found for {distribution.raw_name}."
-            ),
-            hint_stmt=hint,
-        )
-
-
-class LegacyDistutilsInstall(DiagnosticPipError):
-    reference = "uninstall-distutils-installed-package"
-
-    def __init__(self, *, distribution: BaseDistribution) -> None:
-        super().__init__(
-            message=Text(f"Cannot uninstall {distribution}"),
-            context=(
-                "It is a distutils installed project and thus we cannot accurately "
-                "determine which files belong to it which would lead to only a partial "
-                "uninstall."
-            ),
-            hint_stmt=None,
-        )
 
 
 class InvalidInstalledPackage(DiagnosticPipError):
